More rendering fixes/improvements.
multipart/alternative driver now works on basic messages. Be more
liberal about what an 'attachment' is for now. Determination of what
should be shown inline ultimately rests with view code, not base
IMP_Contents library code. Also, it is necessary to determine the
rendered data before we can create the list of attachments (since the
attachments may already be displayed in the body).