Rework download all/strip all links
Move Download All and Strip All links to a separate message action
dropdown since they are not used very often (it also makes the screen
look cleaner). Also add the ability to switch between attachment view
(currently the default) and view all parts (which, for most users, is
normally not useful but every once awhile - especially during
development - it is very useful to be able to view the entire structure
of a message).