Use new rendering code.
getSummary() now only returns a single part info. It is the job of the
calling code to plan the UI output and step through the message
accordingly. This should (eventually) prevent us from having to create
Horde_Mime_Viewer objects for every part of the message, and should be a
bit more memory friendly (since we are only dealing with a part at a
time.