Allow remaining images on DIMP main page to be loaded via Data URL
Loading DIMP's index page on a supported browser, with CSS/JS caching,
now results in 0 images that need to be separately loaded.

Should probably move this to Horde core proper ASAP (along with all
JS/CSS caching code) so that other apps can make use of this. This is
less effective for non-AJAXy applications since the data needs to be
sent on every page load (vs. the image data being cached in CSS).
However, some images are generated on the fly/don't reside in the app
themes directory so they aren't available to be served via CSS.