Improve quicksearch in dimp.
An example of improvement by subtraction - quicksearch in dimp view
should only search entire text of all messages only in the current mailbox.
More advanced search queries need to be implemented in a different UI -
a UI that will be incorporated in the regular imp view also.

Additionally, remove filter interface from ViewPort.  A filter interface
to viewport doesn't make much sense, especially considering the entire
data for the view may not be loaded in the browser.  It is best to deal
with these search views entirely on the server.