Fix escaping null characters in mailbox names.
The null character can not be used in form elements. However, we use the
null character in mailbox names to indicate mailboxes/elements that do
not live on the IMAP server - necessary because null is the only
character technically not allowed for IMAP mailboxes.  So we need to
manually call IMP::formMbox() when creating forms to escape mailbox
names (and then when reading forms to unescape the names - but most of
the time this will be done automatically by
IMP::setCurrentMailboxInfo()).