Queries need to be built for each mailbox
Right now, it is impossible to do an advanced search query - in other
words, a query that involves more than an IMAP query statement.
Creating queries by mailbox allows advanced queries to be performed.
For example, a query element can independently run a query, get the list
of indices that match, and then put this result into the master IMAP
server query with a UID sequence search.