Bug #9912: Fix loading of IMP.* js functions in IE 7
Due to IE 7's inane loading order, it is possible that
IMP.iframeInject() is not available when called.  Instead, refactor
imp.js to store the helper functions in IMP_JS instead of contorting the
code to force this code into the IMP namespace (IMP can't be used until
after javascript_defs.php is loaded, since it is overwritten in there).