Improvements to AJAX framework
Add an _init() method so that apps don't have to extend the more
complicated constructor (with it's 3 arguments).

Add a helper framework to allow AJAX methods to be easily added to an
application.

Move Core AJAX methods to helpers - applications now need to manually
add the helpers they need instead of relying on these "base" methods.