Revert "Set deleted flag after deleting messages (Bug #11362)."
This reverts commit 0f2825e310ac49283d08a35c2d19d199af5452e5.

Can't do this, since it does not correctly update the mailbox cache value.

Regardless, flags should never be set on the browser-side - we need
confirmation from the IMAP server that it has been set on the server or
else it is a usability nightmare (flag setting MUST be an atomic
operation from the POV of the user).