Update Horde_Histories table for IMP_Maillog entries.
Must use ':' as the delimiter in order for change detection to work
sanely. Using a DB migration for the Horde_Histories table from
within IMP is a bit ugly, but because of the History API requiring
this delimiter, it's impossible to perform these changes using either
the IMP_Maillog or the Horde_History API.

The History API should eventually be refactored to build the object_uid
values internally, but that would be impossible before H6.

Placing this in master since I consider this a bug fix, and not a new
feature.