[mms] Rename 'compose_attach' hook as 'compose_attachment'.
Due to the way we build attachment data, the old compose_attach is
worthless.  First, we have moved MIME detection back into IMP's core
code. Also, we need to run the hook for any attachment, not just
uploaded attachment (since we now allow things like browser data URL
insertion).

It is best to simply start all over with a new hook, so thus the total
renaming. Bonus, besides above, is that we expose the actual Attachment
object so ANYTHING about the attachment can be altered (data, MIME part
information) and we still keep the efficiencies of allowing access to
the already existing temporary file.