Rewrite mailbox row (subject) dynamic creation
This is the proper fix for the XSS flag issue fixed in 6.1.8 - this is
the reason you use writeAttribute on an actual DOM element object (rather
than HTML strings) to set attribute values.

We should probably get rid of the message row templating code, since it
isn't really useful - this is not free-form HTML, since there is a bunch
of javascript/CSS to enable this to work correctly as a dynamic table,
as opposed to just a presentation layer.

While in here, fix issue where the quick search critieria is only
updated for the column that is actually being searched on.