Faster operations for conversion
To convert any variable into binary numeric output (i.e. 0 or 1),
a double logical not (!!) will ensure comparison item is boolean, and then
a double bitwise not converts this value to 1 or 0.

Boolean() and Number() is not the preferred way of doing this, according
to various people.  It is also something like 97% slower than doing
bitwise comparisons.