*Finally* get to main fix: replace of changed mailbox instead of delete/add
To do this, we need to keep the state of the previous element in the
object instead of storing within the DOM element itself.

Turns out we already have a data structure that is fully documented re: the
mailbox object -- the AJAX response from IMP_Queue#_flistElt -- so we
should just use that internally.

During this switchover, simplify methods where we don't need set access
outside of the Flist class. Additionally, separate polled status from
unread status.

Add labelElement() return.