[mms] Remove the 'default_msg_charset' preference.
I haven't found a message that displays correctly when specifying a charset
vs. the default '' value.  And it turns out that I've actually been
using '' as my value all along, so messages that display fine for me may
not be displaying properly for others.

We are already doing charset autodetection in Horde_Mime_Headers.  And
browsers do a reasonable job of trying to display the right charset when
given a "default" charset like US-ASCII (vs. something like UTF-8) so
there is no need fo this anymore.

Conflicts:
	imp/package.xml