Fix trying to paste image data in IE 10
A copy/paste of image data in the HTML editor is NOT pasting the
image data; it is instead pasting the HTML source of the image. if the
image is attachment image data, this breaks things (i.e. deleting the
original image attachment will cause the pasted image to also be deleted
in the message, but this won't be shown in the compose window).