Fix undo/redo when adding/removing an image from the HTML compose editor
Can't allow this, since the image data no longer exists on server when
undoing a delete, and the attachment numbering is incorrect on undoing
an add.

Unfortunately, there is no way of disabling/removing the last action so
the only option given to us by the ckeditor API is to clear the undo
stack completely.