Fix GPG Publkey import when also provided by Hook
If a GPG Publickey happens to be provided by a gpg_key Hook in IMP it
was not possible for a user to import said key into his keyring.

This Patch skips gpg_key hook when getPublicKey is called to determine if
a user already has the PublicKey in his keyring.